+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/index_and_scale.fx
@@ -10,7 +10,7 @@ float4 main(float2 texcoord : TEXCOORD) : SV_Target
 	r0 = float4(0, 0, 0, 0);
 	r1.x = 0;
 	while (true) {
-		r1.y = (r1.x >= 4) ? -1 : 0;
+		r1.y = ((uint)r1.x >= 4) ? -1 : 0;
 		if (r1.y != 0) break;
 		r1.y = r1.x * stride;
 		r1.y = r1.y & 3;
