+++ HlslDecompiler.Tests/ShaderSources/vs_4_0/integer_inputs.fx
@@ -19,7 +19,7 @@ VS_OUT main(VS_IN i)
 	VS_OUT o;
 
 	o.sv_position = mul(i.position, wvp);
-	o.color = palette[((uint)i.blendindices.x >> 4) & 3] * 0.00392156886 * (float4)(i.blendindices.y & 255) + (float4)(i.texcoord + 1);
+	o.color = palette[(i.blendindices.x >> 4) & 3] * 0.00392156886 * (float4)(i.blendindices.y & 255) + (float4)(i.texcoord + 1);
 
 	return o;
 }
